third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-css-text.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<style id="style1">
.foo::content div .bar::before { display: block; }
::content { display: block; }
div content::content div content::content div.green { color: green; }
*::content * { color: blue; }
div::shadow::shadow::content { }
</style>
</head>
<script>
description("Test for cssText of '::content' rule.");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(0).cssText", ".foo::content div .bar::before { display: block; }");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(1).cssText", "::content { display: block; }");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(2).cssText", "div content::content div content::content div.green { color: green; }");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(3).cssText", "::content * { color: blue; }");
shouldBe("document.getElementById('style1').sheet.cssRules.length", "4");
</script>
</html>
