third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-css-text-expected.txt

Test for cssText of '::content' rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('style1').sheet.cssRules.item(0).cssText is ".foo::content div .bar::before { display: block; }"
PASS document.getElementById('style1').sheet.cssRules.item(1).cssText is "::content { display: block; }"
PASS document.getElementById('style1').sheet.cssRules.item(2).cssText is "div content::content div content::content div.green { color: green; }"
PASS document.getElementById('style1').sheet.cssRules.item(3).cssText is "::content * { color: blue; }"
PASS document.getElementById('style1').sheet.cssRules.length is 4
PASS successfullyParsed is true

TEST COMPLETE
